First Try to Write in Firebase Realtime DB
I tried to write something to the firebase realtime
database, I am not sure about exactly what happened
as I am still trying things out, I first tried
to follow up the generated answer by chat gpt,
then it did not work when I tried to execute the
js file with node and tsc.

So I tried to link the js file with firebase write
code to the `app.component.html` so that I can test
out if something could be written to the firebase db,
it worked but I don't know how since everytime the
project is served there is no changes to the database.