Followed Tutorial - Failed Tutorial
I tried to follow a simple and short tutorial
about creating a chat app, and it did not work
as expected, all I did was install the getStream.io
angular chat library for creating a chat application
with angular.

The tutorial did not take into account the cases
wherein there would be errors and exception when
following the step-by-step, the details and
credentials of the chat-project is also not available
in the tutorial.